summarize_quiz_generator/requirements.txt
flask>=2.3.0
openai-whisper>=20230918
google-generativeai>=0.3.0
firebase-admin>=6.2.0
